Move new data source to its own space and revert collateral change to ports
