+++ equinix/data_source_fabric_port.go
@@ -7,46 +7,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func readFabricPortResourceSchemaUpdated() map[string]*schema.Schema {
-	sch := FabricPortResourceSchema()
-	sch["uuid"].Computed = true
-	sch["uuid"].Optional = false
-	sch["uuid"].Required = false
-	return sch
-}
-
-func readGetPortsByNameQueryParamSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"name": {
-			Type:        schema.TypeString,
-			Required:    true,
-			Description: "Query Parameter to Get Ports By Name",
-		},
-	}
-}
-
-func readFabricPortsResponseSchema() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"data": {
-			Type:        schema.TypeList,
-			Computed:    true,
-			Description: "List of Ports",
-			Elem: &schema.Resource{
-				Schema: readFabricPortResourceSchemaUpdated(),
-			},
-		},
-		"filters": {
-			Type:        schema.TypeSet,
-			Required:    true,
-			Description: "name",
-			MaxItems:    1,
-			Elem: &schema.Resource{
-				Schema: readGetPortsByNameQueryParamSch(),
-			},
-		},
-	}
-}
-
 func dataSourceFabricPort() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: dataSourceFabricPortRead,
