+++ equinix/resource_fabric_cloud_router.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"github.com/equinix/equinix-sdk-go/services/fabricv4"
-	"github.com/equinix/terraform-provider-equinix/internal/converters"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
@@ -308,135 +307,6 @@ func resourceFabricCloudRouterRead(ctx context.Context, d *schema.ResourceData, 
 	return setCloudRouterMap(d, cloudRouter)
 }
 
-func cloudRouterFiltersTerraformToGo(filters []interface{}) (fabricv4.CloudRouterFilters, error) {
-	if filters == nil || len(filters) == 0 {
-		return fabricv4.CloudRouterFilters{}, fmt.Errorf("no filters passed to filtersTerraformToGoMethod")
-	}
-	cloudRouterFiltersList := make([]fabricv4.CloudRouterFilter, 0)
-	cloudRouterOrFilter := fabricv4.CloudRouterOrFilter{}
-
-	log.Printf("1st filter map %v", filters[0].(map[string]interface{}))
-
-	for _, filter := range filters {
-		filterMap := filter.(map[string]interface{})
-		log.Printf("Filter map %v", filterMap)
-		cloudRouterFilter := fabricv4.CloudRouterFilter{}
-		filterExpression := fabricv4.CloudRouterSimpleExpression{}
-		if property, ok := filterMap["property"]; ok {
-			filterExpression.SetProperty(property.(string))
-		}
-		if operator, ok := filterMap["operator"]; ok {
-			filterExpression.SetOperator(operator.(string))
-		}
-		if values, ok := filterMap["values"]; ok {
-			stringValues := converters.IfArrToStringArr(values.([]interface{}))
-			filterExpression.SetValues(stringValues)
-		}
-
-		// If the parent has any contents then all the children schema properties will be included in the map even
-		// if they aren't given a value. Still need to check for empty string for the value because of this.
-		if orGroup, ok := filterMap["or"]; ok && orGroup.(bool) {
-			orValues := cloudRouterOrFilter.GetOr()
-			orValues = append(orValues, filterExpression)
-			if len(orValues) > 3 {
-				return fabricv4.CloudRouterFilters{}, fmt.Errorf("too many OR group filters passed. Passed %d but can only have a maximum of 3", len(orValues))
-			}
-			cloudRouterOrFilter.SetOr(orValues)
-		} else {
-			cloudRouterFilter.CloudRouterSimpleExpression = &filterExpression
-			cloudRouterFiltersList = append(cloudRouterFiltersList, cloudRouterFilter)
-		}
-	}
-
-	if orGroupHasValues := cloudRouterOrFilter.GetOr(); len(orGroupHasValues) > 0 {
-		cloudRouterFilter := fabricv4.CloudRouterFilter{}
-		cloudRouterFilter.CloudRouterOrFilter = &cloudRouterOrFilter
-		cloudRouterFiltersList = append(cloudRouterFiltersList, cloudRouterFilter)
-	}
-
-	cloudRouterFilters := fabricv4.CloudRouterFilters{}
-	cloudRouterFilters.SetAnd(cloudRouterFiltersList)
-
-	if len(cloudRouterFilters.GetAnd()) > 8 {
-		return fabricv4.CloudRouterFilters{}, fmt.Errorf("too many filters are applied to the data source. The maximum is 8 and %d were provided. Please reduce your filter count to 8", len(cloudRouterFilters.GetAnd()))
-	}
-
-	return cloudRouterFilters, nil
-}
-
-func cloudRouterPaginationTerraformToGo(pagination []interface{}) fabricv4.PaginationRequest {
-	if pagination == nil || len(pagination) == 0 {
-		return fabricv4.PaginationRequest{}
-	}
-	paginationRequest := fabricv4.PaginationRequest{}
-	for _, page := range pagination {
-		pageMap := page.(map[string]interface{})
-		if offset, ok := pageMap["offset"]; ok {
-			paginationRequest.SetOffset(int32(offset.(int)))
-		}
-		if limit, ok := pageMap["limit"]; ok {
-			paginationRequest.SetLimit(int32(limit.(int)))
-		}
-	}
-
-	return paginationRequest
-}
-
-func cloudRouterSortTerraformToGo(sort []interface{}) []fabricv4.CloudRouterSortCriteria {
-	if sort == nil || len(sort) == 0 {
-		return []fabricv4.CloudRouterSortCriteria{}
-	}
-	sortCriteria := make([]fabricv4.CloudRouterSortCriteria, len(sort))
-	for index, item := range sort {
-		sortItem := fabricv4.CloudRouterSortCriteria{}
-		pageMap := item.(map[string]interface{})
-		if direction, ok := pageMap["direction"]; ok {
-			sortItem.SetDirection(fabricv4.CloudRouterSortDirection(direction.(string)))
-		}
-		if property, ok := pageMap["property"]; ok {
-			sortItem.SetProperty(fabricv4.CloudRouterSortBy(property.(string)))
-		}
-		sortCriteria[index] = sortItem
-	}
-	return sortCriteria
-}
-
-func resourceFabricCloudRoutersSearch(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	client := meta.(*config.Config).NewFabricClientForSDK(d)
-	cloudRouterSearchRequest := fabricv4.CloudRouterSearchRequest{}
-
-	schemaFilters := d.Get("filter").([]interface{})
-	filters, err := cloudRouterFiltersTerraformToGo(schemaFilters)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	cloudRouterSearchRequest.SetFilter(filters)
-
-	if schemaPagination, ok := d.GetOk("pagination"); ok {
-		pagination := cloudRouterPaginationTerraformToGo(schemaPagination.(*schema.Set).List())
-		cloudRouterSearchRequest.SetPagination(pagination)
-	}
-
-	if schemaSort, ok := d.GetOk("sort"); ok {
-		sort := cloudRouterSortTerraformToGo(schemaSort.([]interface{}))
-		cloudRouterSearchRequest.SetSort(sort)
-	}
-
-	cloudRouters, _, err := client.CloudRoutersApi.SearchCloudRouters(ctx).CloudRouterSearchRequest(cloudRouterSearchRequest).Execute()
-
-	if err != nil {
-		return diag.FromErr(equinix_errors.FormatFabricError(err))
-	}
-
-	if len(cloudRouters.Data) < 1 {
-		return diag.FromErr(fmt.Errorf("no records are found for the cloud router search criteria provided - %d , please change the search criteria", len(cloudRouters.Data)))
-	}
-
-	d.SetId(cloudRouters.Data[0].GetUuid())
-	return setFabricCloudRoutersData(d, cloudRouters)
-}
-
 func fabricCloudRouterMap(fcr *fabricv4.CloudRouter) map[string]interface{} {
 	package_ := fcr.GetPackage()
 	location := fcr.GetLocation()
@@ -464,25 +334,6 @@ func fabricCloudRouterMap(fcr *fabricv4.CloudRouter) map[string]interface{} {
 		"connections_count":            fcr.GetConnectionsCount(),
 		"order":                        equinix_fabric_schema.OrderGoToTerraform(&order),
 	}
-}
-
-func setFabricCloudRoutersData(d *schema.ResourceData, cloudRouters *fabricv4.SearchResponse) diag.Diagnostics {
-	diags := diag.Diagnostics{}
-	mappedCloudRouters := make([]map[string]interface{}, len(cloudRouters.Data))
-	if cloudRouters.Data != nil {
-		for index, cloudRouter := range cloudRouters.Data {
-			mappedCloudRouters[index] = fabricCloudRouterMap(&cloudRouter)
-		}
-	} else {
-		mappedCloudRouters = nil
-	}
-	err := equinix_schema.SetMap(d, map[string]interface{}{
-		"data": mappedCloudRouters,
-	})
-	if err != nil {
-		return diag.FromErr(err)
-	}
-	return diags
 }
 
 func setCloudRouterMap(d *schema.ResourceData, fcr *fabricv4.CloudRouter) diag.Diagnostics {
